WICKET-6287 Switch from json.org to open-json - 8.x
Repository: https://github.com/tdunning/open-json
