+++ pom.xml
@@ -447,6 +447,12 @@
 				<type>jar</type>
 			</dependency>
 			<dependency>
+		    	<groupId>com.tdunning</groupId>
+		    	<artifactId>json</artifactId>
+				<version>1.2-SNAPSHOT</version>
+				<type>jar</type>
+			</dependency>
+			<dependency>
 				<groupId>org.jboss.seam.conversation</groupId>
 				<artifactId>seam-conversation-spi</artifactId>
 				<version>3.0.0.Final</version>
