+++ wicket-core/pom.xml
@@ -66,6 +66,10 @@
 			<groupId>org.mockito</groupId>
 			<artifactId>mockito-core</artifactId>
 		</dependency>
+		<dependency>
+		    <groupId>com.tdunning</groupId>
+		    <artifactId>json</artifactId>
+		</dependency>
 	</dependencies>
 	<build>
 		<pluginManagement>
