--- wicket-core/src/main/java/org/apache/wicket/ajax/json/JSONException.java
@@ -1,41 +1,0 @@
-package org.apache.wicket.ajax.json;
-
-/**
- * The JSONException is thrown by the JSON.org classes when things are amiss.
- *
- * @author JSON.org
- * @version 2013-02-10
- */
-public class JSONException extends RuntimeException {
-    private static final long serialVersionUID = 0;
-    private Throwable cause;
-
-    /**
-     * Constructs a JSONException with an explanatory message.
-     *
-     * @param message
-     *            Detail about the reason for the exception.
-     */
-    public JSONException(String message) {
-        super(message);
-    }
-
-    /**
-     * Constructs a new JSONException with the specified cause.
-     */
-    public JSONException(Throwable cause) {
-        super(cause.getMessage());
-        this.cause = cause;
-    }
-
-    /**
-     * Returns the cause of this exception or null if the cause is nonexistent
-     * or unknown.
-     *
-     * @returns the cause of this exception or null if the cause is nonexistent
-     *          or unknown.
-     */
-    public Throwable getCause() {
-        return this.cause;
-    }
-}
