+++ wicket-core/src/main/java/org/apache/wicket/ajax/json/JsonUtils.java
@@ -19,6 +19,10 @@ package org.apache.wicket.ajax.json;
 import java.util.List;
 import java.util.Map;
 
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
 /**
  * @since 6.0.0
  */
